+++ src/gitversioniser/domain/repository/commit.py
@@ -4,26 +4,15 @@ from git.objects import Commit as GitCommit
 
 from gitversioniser.config.config import Config
 from gitversioniser.domain.repository.commit_utils.message import Message
-from gitversioniser.domain.repository.commit_utils.summary import Summary
 
 
 @dataclass
 class Commit:
-    """ Couldn't make it with inheritance. """
     config: Config = field(repr=False)
     _commit: GitCommit = field(repr=False)
 
     def __post_init__(self):
         self.message = Message(self.config, str(self._commit.message))
-
-    @property
-    def summary(self) -> Summary:
-        """ Summary is just a part of message, but it's handy to use it as a property. """
-        return self.message.summary
-
-    @property
-    def description(self) -> str:
-        return self.message.value.replace(str(self.message.summary), '')
 
     @property
     def get_parents_count(self) -> int: